requirements-dev.txt
# SutazAI Development Requirements
# Generated: 2025-08-07
# Development tools, linting, formatting, and debugging utilities

# Include production requirements
-r requirements.txt

# ============================================================================
# CODE QUALITY & LINTING
# ============================================================================
black==24.10.0                     # Code formatter - latest stable
ruff==0.8.6                        # Fast Python linter - latest features
mypy==1.14.1                       # Static type checker - latest
isort==5.13.2                      # Import sorting
flake8==6.1.0                      # Legacy linter (for compatibility)
bandit==1.7.10                     # Security linter - latest security checks
safety==3.2.11                     # Dependency vulnerability scanner
pip-audit==2.7.3                   # Additional security scanning

# ============================================================================
# PRE-COMMIT HOOKS
# ============================================================================
pre-commit==4.0.1                  # Pre-commit framework - latest

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================
ipython==8.29.0                    # Enhanced Python REPL
jupyter==1.1.1                     # Jupyter meta-package
notebook==7.3.1                    # Jupyter notebook
jupyterlab==4.3.3                  # JupyterLab IDE
ipdb==0.13.13                      # Debugger
pdb-attach==3.2.0                  # Remote debugging
python-dotenv[cli]==1.0.1          # CLI for .env management

# ============================================================================
# DOCUMENTATION GENERATION
# ============================================================================
sphinx==8.1.3                      # Documentation generator - latest
sphinx-rtd-theme==3.0.2            # ReadTheDocs theme
myst-parser==4.0.0                 # Markdown support for Sphinx
sphinx-autodoc-typehints==2.4.4    # Type hints in docs
mkdocs==1.6.1                      # Alternative docs generator
mkdocs-material==9.5.44            # Material theme for MkDocs

# ============================================================================
# PROFILING & PERFORMANCE
# ============================================================================
line-profiler==4.1.3               # Line-by-line profiling
memory-profiler==0.61.0            # Memory usage profiling
pympler==0.9                       # Advanced memory profiling
objgraph==3.6.2                    # Object reference tracking
py-spy==0.3.14                     # Sampling profiler
scalene==1.5.47                    # High-performance CPU/memory profiler

# ============================================================================
# DATABASE DEVELOPMENT
# ============================================================================
pgcli==4.1.0                       # PostgreSQL CLI with autocomplete
mycli==1.27.2                      # MySQL CLI (if needed)
redis-cli==2.0.2                   # Redis CLI improvements

# ============================================================================
# API DEVELOPMENT & TESTING
# ============================================================================
httpie==3.2.4                      # Command-line HTTP client
postman-cli==1.0.0                 # Postman CLI (if available)
insomnia-cli==1.0.0                # Insomnia CLI (if available)

# ============================================================================
# CONTAINERS & ORCHESTRATION
# ============================================================================
docker-compose==1.29.2             # Docker Compose (if not using v2)
kubernetes==31.0.0                 # Kubernetes client - security validated
helm==3.14.0                       # Helm package manager (if available)

# ============================================================================
# MONITORING & DEBUGGING
# ============================================================================
watchdog==6.0.0                    # File system monitoring
python-json-logger==2.0.7          # JSON logging for development
colorlog==6.9.0                    # Colored logging output
loguru==0.7.2                      # Advanced logging (dev features)

# ============================================================================
# TESTING ENHANCEMENTS (Extended from test requirements)
# ============================================================================
pytest-sugar==1.0.0                # Beautiful test output
pytest-clarity==1.0.1              # Clear assertion diffs
pytest-picked==0.5.0               # Run tests from modified files
pytest-watch==4.2.0                # Continuous testing
pytest-testmon==2.1.3              # Smart test selection
pytest-benchmark==4.0.0            # Performance benchmarking

# ============================================================================
# GIT & VERSION CONTROL
# ============================================================================
gitpython==3.1.43                  # Git interface for Python
commitizen==3.29.1                 # Semantic versioning/changelog
bump2version==1.0.1                # Version bumping utility

# ============================================================================
# ENVIRONMENT MANAGEMENT
# ============================================================================
pipenv==2024.0.3                   # Environment management (if used)
virtualenv==20.27.1                # Virtual environment creation
virtualenvwrapper==6.1.0           # Virtualenv extensions

# ============================================================================
# CODE ANALYSIS & COMPLEXITY
# ============================================================================
radon==6.0.1                       # Code complexity analysis
vulture==2.13                      # Dead code finder
mccabe==0.7.0                      # McCabe complexity checker
lizard==1.17.10                    # Code complexity analyzer

# ============================================================================
# JUPYTER EXTENSIONS
# ============================================================================
jupyterlab-git==0.50.2             # Git extension for JupyterLab
nbstripout==0.7.1                  # Clean Jupyter notebooks
jupytext==1.16.6                   # Jupyter notebook text formats

# ============================================================================
# DEVELOPMENT UTILITIES
# ============================================================================
faker==33.1.0                      # Fake data generation
factory-boy==3.3.1                 # Test data factories
freezegun==1.5.1                   # Time mocking
responses==0.25.3                  # HTTP request mocking
requests-mock==1.12.1              # Request mocking

# ============================================================================
# DEVELOPMENT CONFIGURATION
# ============================================================================
python-decouple==3.8               # Configuration management
pydantic[dotenv]==2.10.4          # Enhanced .env support
dynaconf==3.2.6                    # Advanced configuration

# ============================================================================
# AI/ML DEVELOPMENT
# ============================================================================
wandb==0.19.0                      # ML experiment tracking
tensorboard==2.18.0                # TensorFlow visualization
mlflow==2.19.0                     # ML lifecycle management